refactor: extract helper functions from SelectAreaLayer to utils
